Move api/server/router/container to daemon/server/router/container
